Moves flag-unlabeled-after-deletion scripts out of sub-directory
