+++ github-actions/flag-issues-unlabeled-after-deletion/format-issue-list.js
